pyproject.toml
[project]
name = "medscope-server"
version = "0.1.0"
description = ""
authors = [
    {name = "unfortunatelygeek",email = "raoaditisudhir@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "albumentations (>=2.0.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
